wrap data into Engine struct + store txs
I'll have to store transactions along with accounts because when
transactions are disputed the input references a transaction by its ID
rather than providing the amount to dispute, so I have to do a lookup.

I'm also wrapping up the transaction and account maps into an `Engine`
struct mainly to keep things a little more tidy (at least for now). I've
updated the test to use `Engine` instead of the `AcctMngr` from the last
commit.