implement disputes + start custom errors
The transaction map is now a new transaction type instead of the parsed
CSV ones, which holds only the necessary information I need for
disputing transactions.

Also added more details in the README about how disputes are handled
with some examples of how account information changes with the different
stages of disputes.

I've added a diagram to show the dispute state diagram as well.