input csv generator + fixed bug
Created a simple CSV generator in a Python 3 script. This actually
helped me find a bug where I wasn't checking that a dispute-related
transaction ID was actually associated with the correct client. Added a
new test to cover this case.

Also, I'm now capturing the errors in the `TestDef::run` function
instead of returning an error which is more like the actual program.